+++ src/main/java/seedu/address/model/delivery/Date.java
@@ -1,4 +1,4 @@
-package seedu.address.model.date;
+package seedu.address.model.delivery;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
